修复商品加载，数据解析失败问题；优化货品图片显示，解决部分图片显示变形问题

+++ app/src/main/java/com/yifarj/yifadinghuobao/model/entity/GoodsListEntity.java
@@ -40,7 +40,7 @@ public class GoodsListEntity {
         public int DefaultTraderId;
         public int PurchaserId;
         public int ManagerId;
-        public int Cost;
+        public double Cost;
         public double MinSalesQuantity;
         public double MaxSalesQuantity;
         public int CostCalcMethodId;
